requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
django-filter==21.1
django==3.2
djangorestframework-simplejwt==5.0
djangorestframework==3.12
markdown==3.3
psycopg2-binary==2.9
pyjwt==2.3.0; python_version >= '3.6'
pytz==2021.3
sqlparse==0.4.2; python_version >= '3.5'
